+++ Application/Salvation.Core/Constants/ConduitData.cs
@@ -7,12 +7,12 @@ namespace Salvation.Core.Constants
         public int Id { get; set; }
         public string Name { get; set; }
         public int ConduitId { get; set; }
-        public List<decimal> Ranks { get; set; }
-        public decimal Coeff1 { get; set; }
+        public List<double> Ranks { get; set; }
+        public double Coeff1 { get; set; }
 
         public ConduitData()
         {
-            Ranks = new List<decimal>();
+            Ranks = new List<double>();
         }
     }
 }
